refactor: export leaf_index_to_pos
