+++ util/merkle-mountain-range/src/lib.rs
@@ -1,4 +1,5 @@
 mod error;
+mod helper;
 mod merkle_elem;
 mod mmr;
 mod mmr_store;
@@ -7,6 +8,7 @@ mod tests;
 pub mod tests_util;
 
 pub use error::{Error, Result};
+pub use helper::leaf_index_to_pos;
 pub use merkle_elem::MerkleElem;
 pub use mmr::{MerkleProof, MMR};
 pub use mmr_store::MMRStore;
