+++ util/merkle-mountain-range/src/mmr.rs
@@ -5,6 +5,7 @@
 //! https://github.com/mimblewimble/grin/blob/0ff6763ee64e5a14e70ddd4642b99789a1648a32/core/src/core/pmmr.rs#L606
 
 use crate::error::Result;
+use crate::helper::{get_peaks, parent_offset, pos_height_in_tree, sibling_offset};
 use crate::mmr_store::MMRStore;
 use crate::MerkleElem;
 use ckb_db::KeyValueDB;
@@ -17,78 +18,6 @@ pub struct MMR<Elem, DB> {
     mmr_size: u64,
     store: Arc<MMRStore<Elem, DB>>,
     merkle_elem: PhantomData<Elem>,
-}
-
-fn pos_height_in_tree(mut pos: u64) -> u64 {
-    pos += 1;
-    fn all_ones(num: u64) -> bool {
-        num.count_zeros() == num.leading_zeros()
-    }
-    fn jump_left(pos: u64) -> u64 {
-        let bit_length = 64 - pos.leading_zeros();
-        let most_significant_bits = 1 << (bit_length - 1);
-        pos - (most_significant_bits - 1)
-    }
-
-    while !all_ones(pos) {
-        pos = jump_left(pos)
-    }
-
-    (64 - pos.leading_zeros() - 1).into()
-}
-
-fn parent_offset(height: u32) -> u64 {
-    2 << height
-}
-
-fn sibling_offset(height: u32) -> u64 {
-    (2 << height) - 1
-}
-
-fn get_peaks(mmr_size: u64) -> Vec<u64> {
-    let mut pos_s = Vec::new();
-    let (mut height, mut pos) = left_peak_height_pos(mmr_size);
-    pos_s.push(pos);
-    while height > 0 {
-        let peak = match get_right_peak(height, pos, mmr_size) {
-            Some(peak) => peak,
-            None => break,
-        };
-        height = peak.0;
-        pos = peak.1;
-        pos_s.push(pos);
-    }
-    pos_s
-}
-
-fn get_right_peak(mut height: u32, mut pos: u64, mmr_size: u64) -> Option<(u32, u64)> {
-    // move to right sibling pos
-    pos += sibling_offset(height);
-    // loop until we find a pos in mmr
-    while pos > mmr_size - 1 {
-        if height == 0 {
-            return None;
-        }
-        // move to left child
-        pos -= parent_offset(height - 1);
-        height -= 1;
-    }
-    Some((height, pos))
-}
-
-fn left_peak_height_pos(mmr_size: u64) -> (u32, u64) {
-    fn get_left_pos(height: u32) -> u64 {
-        (1 << (height + 1)) - 2
-    }
-    let mut height = 0;
-    let mut prev_pos = 0;
-    let mut pos = get_left_pos(height);
-    while pos < mmr_size {
-        height += 1;
-        prev_pos = pos;
-        pos = get_left_pos(height);
-    }
-    (height - 1, prev_pos)
 }
 
 impl<Elem: MerkleElem + Clone + Eq + Debug, DB: KeyValueDB> MMR<Elem, DB> {
